pyproject.toml
[tool.poetry]
name = "fastapi-crud"
version = "0.1.0"
description = ""
authors = ["eventer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["all"], version = "^0.111.0" }
sqlmodel = "^0.0.19"
alembic = "^1.13.2"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
cryptography = "^42.0.8"
pydantic-settings = "^2.3.4"
uvicorn = "^0.30.1"
pymysql = "^1.1.1"
pyjwt = "^2.8.0"

[tool.poetry.scripts]
migrate = "command:migrate"
dev = "command:dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
